% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReports.R
\name{addIdentifiers}
\alias{addIdentifiers}
\title{Given Demuxed data, add in sample barcode information}
\usage{
addIdentifiers(...)
}
\description{
If cfg environment variable is set, this function will merge sample ID info onto amplicon count tables.
}
