% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPullSeq.R
\name{autoRun}
\alias{autoRun}
\title{Auto Run the Pipeline}
\usage{
autoRun(check.yaml.interval = 30, check.basespace.interval = 600)
}
\arguments{
\item{check.yaml.interval}{seconds before checking config.yaml file for updates}

\item{check.basespace.interval}{seconds before checking basespace to see if a run has updated}
}
\description{
If cfg environment variable is set, this function will autorun the pipeline.
}
