% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnvironment.R
\name{buildEnvironment}
\alias{buildEnvironment}
\title{Build SwabSeq environment variable}
\usage{
buildEnvironment(
  basedir.dir,
  remote.dir,
  localmirror.dir,
  bcl.dir,
  threads = 8,
  lbuffer = 3e+07,
  fastqcr = F
)
}
\description{
This function creates an env() containing information about directory structures and key run parameters.
}
