% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPullSeq.R
\name{demuxRuns}
\alias{demuxRuns}
\title{Run BCl2FASTQ}
\usage{
demuxRuns(...)
}
\description{
If cfg environment variable is set, this function will demux amplicons given key files in $basedir.dir/reference/.
}
