% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPullSeq.R
\name{getRunTableStatus}
\alias{getRunTableStatus}
\title{Read config.yaml and convert to table}
\usage{
getRunTableStatus(...)
}
\description{
If cfg environment variable is set, this function will read the config.yaml table for downstream analysis.
}
