% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPullSeq.R
\name{lookUpKeys}
\alias{lookUpKeys}
\title{Lookup experiments and flowcells and merge with basespace information}
\usage{
lookUpKeys(...)
}
\description{
If cfg environment variable is set, this function will sync basespace runs with the experiment folder, and
sample keys from scanning matrix tubes. Merge is performed based on flowcell file in each run directory.
}
