% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnvironment.R
\name{setAnalysisVariables}
\alias{setAnalysisVariables}
\title{Set Analysis Variables}
\usage{
setAnalysisVariables(
  versi = 2,
  diversifiedSpike = T,
  lbuffer = 3e+07,
  threads = 8,
  fastqcr = F
)
}
\description{
This function is called by buildEnvironment() and sets key analysis parameters, exposed here for user overwriting.
}
