% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCSVReports.R
\name{syncReports}
\alias{syncReports}
\title{Sync Reports to shared drive}
\usage{
syncReports(..., syncToShared = T)
}
\arguments{
\item{syncToShared}{flag to sync results to sample tracking folder, set to F for debugging}
}
\description{
If cfg environment variable is set, this function will generate csv reports and sync to the shared drive
}
