% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPullSeq.R
\name{syncRuns}
\alias{syncRuns}
\title{Sync config.yaml file}
\usage{
syncRuns(first.run = F)
}
\description{
If cfg environment variable is set, this function will sync config.yaml with updates from basespace
}
